+++ app/src/main/java/com/android/shaftschematic/ui/nav/InternalDocRoutes.kt
@@ -24,6 +24,7 @@ import androidx.navigation.NavController
 import com.android.shaftschematic.io.InternalStorage
 import com.android.shaftschematic.model.ShaftPosition
 import com.android.shaftschematic.ui.viewmodel.ShaftViewModel
+import com.android.shaftschematic.ui.viewmodel.unlockAchievement
 import com.android.shaftschematic.util.FeedbackIntentFactory
 import com.android.shaftschematic.util.Achievements
 import com.android.shaftschematic.util.DocumentNaming