+++ app/src/main/java/com/android/shaftschematic/ui/screen/SettingsRoute.kt
@@ -61,6 +61,7 @@ import kotlin.math.roundToInt
 import com.android.shaftschematic.BuildConfig
 import com.android.shaftschematic.ui.viewmodel.ShaftViewModel
 import com.android.shaftschematic.ui.viewmodel.UiEvent
+import com.android.shaftschematic.ui.viewmodel.*
 import com.android.shaftschematic.util.PreviewColorPreset
 import com.android.shaftschematic.util.PreviewColorRole
 import com.android.shaftschematic.util.PreviewColorSetting